[refactor] 도메인 기반 폴더링 구조로 수정
